create ergomath layout
layout where numlock is a "math mode" with digits, abcdef, and most programming symbols are in direct access.
This is the only layout I use now, other will be removed soon (except if someone asks me not to do).
ergomath layout is made for Ergodox keyboard. It will renamed.
Need changes on "level5" file to be run correctly.